Needs requests impl

+++ inventory.pl/src/main/java/inventory/pl/dao/RequestDetailsRepository.java
@@ -13,6 +13,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  *
  * @author ahmed_darweeesh
  */
-public interface RequestDetailsRepository extends JpaRepository<RequestDetails, Long>{
+public interface RequestDetailsRepository extends JpaRepository<RequestDetails, Integer>{
     
 }
